refactor(order): convert monetary fields to integers and update related logic
- Change discountAmount and totalAmount fields to integers in schema and database metadata
- Update order creation logic to floor discountAmount and totalAmount values before storing
- Modify PDF service to display floored totalAmount without decimals
- Adjust WhatsApp message formatting to use integer totalAmount without decimal points
- Update input validation to transform totalAmount and discountAmount strings into integers
- Remove old database migration that created decimal fields tables to avoid conflicts
- Ensure finalTotal and discountAmount are never negative and properly floored before insertion and update
- Harmonize client and server code to consistently handle monetary amounts as integer values

+++ migrations/0000_overconfident_bill_hollister.sql
@@ -59,8 +59,8 @@ CREATE TABLE `orders` (
 	`notes` text,
 	`sub_total` decimal(10,2),
 	`discount_percentage` decimal(5,2),
-	`discount_amount` decimal(10,2),
-	`total_amount` decimal(10,2) NOT NULL,
+	`discount_amount` int,
+	`total_amount` int NOT NULL,
 	`date` timestamp DEFAULT CURRENT_TIMESTAMP,
 	CONSTRAINT `orders_id` PRIMARY KEY(`id`)
 );
